+++ src/audio/codec/mod.rs
@@ -7,3 +7,8 @@ pub use ak4556::{Codec, Pins};
 mod wm8731;
 #[cfg(feature = "seed_1_1")]
 pub use wm8731::{Codec, Pins};
+
+#[cfg(feature = "patch_sm")]
+mod pcm3060;
+#[cfg(feature = "patch_sm")]
+pub use pcm3060::{Codec, Pins};