+++ src/pins.rs
@@ -2,6 +2,7 @@ use stm32h7xx_hal as hal;
 
 // - types --------------------------------------------------------------------
 
+// https://github.com/electro-smith/DaisyWiki/wiki/2.-Daisy-Seed-Pinout
 #[cfg(any(feature = "seed", feature = "seed_1_1"))]
 #[allow(non_snake_case)]
 pub struct Gpio {
@@ -38,6 +39,43 @@ pub struct Gpio {
     pub PIN_30: hal::gpio::gpiob::PB15<hal::gpio::Analog>, // STM PIN 37, USB1 D+, USART1 Rx
 }
 
+// ES_Patch_SM_datasheet_v1.0.3.pdf
+// * 47K pullups are connected to this pin. This may have an effect on the behavior when used as UART
+#[cfg(feature = "patch_sm")]
+#[allow(non_snake_case)]
+pub struct Gpio {
+    pub PIN_A2: hal::gpio::gpioa::PA1<hal::gpio::Analog>, // ADC_9, GPIO, UART4_RX
+    pub PIN_A3: hal::gpio::gpioa::PA0<hal::gpio::Analog>, // ADC_10, GPIO, UART4_TX
+    pub PIN_A8: hal::gpio::gpiob::PB14<hal::gpio::Analog>, // USB_DM, GPIO, USART1_TX
+    pub PIN_A9: hal::gpio::gpiob::PB15<hal::gpio::Analog>, // USB_DP, GPIO, USART1_RX
+    pub PIN_B5: hal::gpio::gpioc::PC14<hal::gpio::Analog>, // GATE_OUT_1 Output Only
+    pub PIN_B6: hal::gpio::gpioc::PC13<hal::gpio::Analog>, // GATE_OUT_2 Output Only
+    pub PIN_B7: hal::gpio::gpiob::PB8<hal::gpio::Analog>, // I2C1_SCL, GPIO, UART4_RX, PWM (TIM4_CH3)
+    pub PIN_B8: hal::gpio::gpiob::PB9<hal::gpio::Analog>, // I2C1_SDA, GPIO, UART4_TX, PWM (TIM4_CH4)
+    pub PIN_B9: hal::gpio::gpiog::PG14<hal::gpio::Analog>, // GATE_IN_2, Input Only
+    pub PIN_B10: hal::gpio::gpiog::PG13<hal::gpio::Analog>, // GATE_IN_1, Input Only
+    pub PIN_C1: hal::gpio::gpioa::PA5<hal::gpio::Analog>, // CV_OUT_2, Output Only
+    pub PIN_C2: hal::gpio::gpioa::PA7<hal::gpio::Analog>, // CV_4, Input Only
+    pub PIN_C3: hal::gpio::gpioa::PA2<hal::gpio::Analog>, // CV_3, Input Only
+    pub PIN_C4: hal::gpio::gpioa::PA6<hal::gpio::Analog>, // CV_2, Input Only
+    pub PIN_C5: hal::gpio::gpioa::PA3<hal::gpio::Analog>, // CV_1, Input Only
+    pub PIN_C6: hal::gpio::gpiob::PB1<hal::gpio::Analog>, // CV_5, Input Only
+    pub PIN_C7: hal::gpio::gpioc::PC4<hal::gpio::Analog>, // CV_6, Input Only
+    pub PIN_C8: hal::gpio::gpioc::PC0<hal::gpio::Analog>, // CV_7, Input Only
+    pub PIN_C9: hal::gpio::gpioc::PC1<hal::gpio::Analog>, // CV_8, Input Only
+    pub PIN_C10: hal::gpio::gpioa::PA4<hal::gpio::Analog>, // CV_OUT_1, Output Only
+    pub PIN_D1: hal::gpio::gpiob::PB4<hal::gpio::Alternate<0>>, // SPI2_CS, GPIO,
+    pub PIN_D2: hal::gpio::gpioc::PC11<hal::gpio::Analog>, // SDMMC1_D3, GPIO, USART3_RX*
+    pub PIN_D3: hal::gpio::gpioc::PC10<hal::gpio::Analog>, // SDMMC1_D2, GPIO, USART3_TX*
+    pub PIN_D4: hal::gpio::gpioc::PC9<hal::gpio::Analog>, // SDMMC1_D1, GPIO
+    pub PIN_D5: hal::gpio::gpioc::PC8<hal::gpio::Analog>, // SDMMC1_D0, GPIO
+    pub PIN_D6: hal::gpio::gpioc::PC12<hal::gpio::Analog>, // SDMMC1_CLK, GPIO, UART5_TX*
+    pub PIN_D7: hal::gpio::gpiod::PD2<hal::gpio::Analog>, // SDMMC1_CMD, GPIO, UART5_RX*
+    pub PIN_D8: hal::gpio::gpioc::PC2<hal::gpio::Analog>, // ADC_12, GPIO, SPI2_MISO
+    pub PIN_D9: hal::gpio::gpioc::PC3<hal::gpio::Analog>, // ADC_11, GPIO, SPI2_MOSI
+    pub PIN_D10: hal::gpio::gpiog::PG3<hal::gpio::Analog>, // SPI2_SCK, GPIO
+}
+
 pub type LedUserPin = hal::gpio::gpioc::PC7<hal::gpio::Analog>; // LED_USER
 
 #[cfg(feature = "seed")]
@@ -51,6 +89,13 @@ pub struct CodecPins {
 pub struct CodecPins {
     pub SCL: hal::gpio::gpioh::PH4<hal::gpio::Analog>, // I2C2 SCL (WM8731)
     pub SDA: hal::gpio::gpiob::PB11<hal::gpio::Analog>, // I2C2 SDA (WM8731)
+}
+
+#[cfg(feature = "patch_sm")]
+#[allow(non_snake_case)]
+pub struct CodecPins {
+    pub SCL: hal::gpio::gpiob::PB10<hal::gpio::Analog>, // I2C2 SCL (PCM3060)
+    pub SDA: hal::gpio::gpiob::PB11<hal::gpio::Analog>, // I2C2 SDA (PCM3060)
 }
 
 #[allow(non_snake_case)]
@@ -83,7 +128,6 @@ pub struct FMCPins {
 
 #[allow(non_snake_case)]
 pub struct Pins {
-    // https://github.com/electro-smith/DaisyWiki/wiki/2.-Daisy-Seed-Pinout
     pub GPIO: Gpio,
 
     // board peripherals