+++ CHANGELOG.md
@@ -6,10 +6,11 @@ backwards compatibility.
 
 ## Unreleased
 
-* Remove unneeded `cty` dependency.
-* Rename `seed_1_0` feature to `seed`.
+* Introduce support for Daisy Patch SM, with PCM3060 codec.
 * Move general purpose pins under a sub-structure. e.g. `SEED_PIN_1` to
   `GPIO.PIN_1`.
+* Rename `seed_1_0` feature to `seed`.
+* Remove unneeded `cty` dependency.
 
 ## 0.2.0